Cambios en la base de datos y solucionado bug de las abreviaturas

+++ src/controllers/CentroController.py
@@ -12,11 +12,12 @@ def index():
 def add():
     formulario = FormCentro()
     if formulario.validate_on_submit():
+        codigo = formulario.codigo.data
         nombre = formulario.nombre.data
         abreviatura = formulario.abreviatura.data
         email = formulario.email.data
 
-        centro = Centro(nombre=nombre, abreviatura=abreviatura, email=email)
+        centro = Centro(codigo=codigo, nombre=nombre, abreviatura=abreviatura, email=email)
         centro.save()
         flash('Centro añadido correctamente', 'alert alert-success alert-dismissible fade show')
         return redirect(url_for('centro_bp.index'))
@@ -30,6 +31,7 @@ def update(id_centro):
     formulario = FormCentro(obj=centro)
     formulario.submit.label.text = 'Modificar'
     if formulario.validate_on_submit():
+        centro.codigo = formulario.codigo.data
         centro.nombre = formulario.nombre.data
         centro.abreviatura = formulario.abreviatura.data
         centro.email = formulario.email.data
@@ -43,6 +45,10 @@ def delete(id_centro):
     centro = Centro.get_centro(id_centro)
     if centro is None:
         abort(404)
+    if centro.titulaciones:
+        flash('No se puede eliminar el centro porque tiene titulaciones asociadas',
+              'alert alert-danger alert-dismissible fade show')
+        return redirect(url_for('centro_bp.index'))
     centro.delete()
     flash('Centro eliminado correctamente', 'alert alert-success alert-dismissible fade show')
     return redirect(url_for('centro_bp.index'))
